Bug 1913350: Use BigInt for memory64 params/results. r=rhunt
This patch does a sweep over the wasm JS API to return BigInt for
relevant memory64 params and results. More precisely, it brings the JS
API in line with the new IndexValue stuff in the JS API spec.

It also fixes validation of table types to allow the larger maximum size
of tables allowed by table64. This changes the limits type of tables to
use uint64_t instead of uint32_t, but does not change any actual
behavior and does not change the internal storage of tables.

Differential Revision: https://phabricator.services.mozilla.com/D219530

+++ js/src/wasm/WasmCodegenTypes.h
@@ -985,8 +985,8 @@ class CalleeDesc {
     } import;
     struct {
       uint32_t instanceDataOffset_;
-      uint32_t minLength_;
-      mozilla::Maybe<uint32_t> maxLength_;
+      uint64_t minLength_;
+      mozilla::Maybe<uint64_t> maxLength_;
       CallIndirectId callIndirectId_;
     } table;
     SymbolicAddress builtin_;
